# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type RateTableRates = {
  one_bedroom: number;
  two_bedroom: number;
};

export interface RateTable {
  currency: string;
  minor_per_major: number;
  rates: RateTableRates;
}

export type GroupOrderInputMode = typeof GroupOrderInputMode[keyof typeof GroupOrderInputMode];


export const GroupOrderInputMode = {
  independent: 'independent',
  split: 'split',
} as const;

export type SplitConfigApartmentType = typeof SplitConfigApartmentType[keyof typeof SplitConfigApartmentType];


export const SplitConfigApartmentType = {
  one_bedroom: 'one_bedroom',
  two_bedroom: 'two_bedroom',
} as const;

export interface SplitConfig {
  apartment_type: SplitConfigApartmentType;
  /** @minimum 1 */
  nights: number;
}

/**
 * Required in independent mode
 * @nullable
 */
export type LineInputApartmentType = typeof LineInputApartmentType[keyof typeof LineInputApartmentType] | null;


export const LineInputApartmentType = {
  one_bedroom: 'one_bedroom',
  two_bedroom: 'two_bedroom',
} as const;

export interface LineInput {
  /** @minLength 1 */
  payer_name: string;
  /** @minLength 1 */
  payer_email: string;
  /**
     * Required in independent mode
     * @nullable
     */
  apartment_type?: LineInputApartmentType;
  /**
     * Required in independent mode
     * @nullable
     */
  nights?: number | null;
  /**
     * Optional explicit share for split mode (minor units)
     * @nullable
     */
  share_minor?: number | null;
}

export interface GroupOrderInput {
  mode: GroupOrderInputMode;
  /** @minLength 1 */
  organiser_name: string;
  /** @minLength 1 */
  organiser_email: string;
  /**
     * ISO 8601 pay-by deadline
     * @nullable
     */
  due_by?: string | null;
  split?: SplitConfig | null;
  lines: LineInput[];
}

export interface CreatedLine {
  id: string;
  payer_name: string;
  payer_email: string;
  amount_minor: number;
  pay_link: string;
}

export interface GroupOrderCreated {
  order_id: string;
  status_token: string;
  organiser_url: string;
  lines: CreatedLine[];
}

export interface OrganiserLine {
  id: string;
  payer_name: string;
  payer_email: string;
  amount_major: number;
  status: string;
  /** @nullable */
  voucher_code?: string | null;
  /** @nullable */
  credit_code?: string | null;
  /** @nullable */
  apartment_type?: string | null;
  /** @nullable */
  nights?: number | null;
  pay_link: string;
}

export interface OrganiserView {
  mode: string;
  status: string;
  currency: string;
  organiser_name: string;
  /** @nullable */
  due_by?: string | null;
  /** @nullable */
  split_apartment_type?: string | null;
  /** @nullable */
  split_nights?: number | null;
  /** @nullable */
  split_voucher_code?: string | null;
  paid_count: number;
  total_count: number;
  lines: OrganiserLine[];
}

export interface ResendResult {
  pay_link: string;
  payer_email: string;
  status: string;
}

export interface PayLine {
  payer_name: string;
  amount_major: number;
  currency: string;
  description: string;
  status: string;
  payable: boolean;
  /** @nullable */
  voucher_code?: string | null;
  organiser_name?: string;
}

export interface CheckoutSession {
  url: string;
}

export interface SweepResult {
  swept: number;
}