# lib/api-zod/src/generated/types/createdLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatedLine {
  id: string;
  payer_name: string;
  payer_email: string;
  amount_minor: number;
  pay_link: string;
}