# lib/api-zod/src/generated/types/groupOrderCreated.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreatedLine } from './createdLine';

export interface GroupOrderCreated {
  order_id: string;
  status_token: string;
  organiser_url: string;
  lines: CreatedLine[];
}