# lib/api-zod/src/generated/types/groupOrderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroupOrderInputMode } from './groupOrderInputMode';
import type { LineInput } from './lineInput';
import type { SplitConfig } from './splitConfig';

export interface GroupOrderInput {
  mode: GroupOrderInputMode;
  /** @minLength 1 */
  organiser_name: string;
  /** @minLength 1 */
  organiser_email: string;
  /**
     * ISO 8601 pay-by deadline
     * @nullable
     */
  due_by?: string | null;
  split?: SplitConfig | null;
  lines: LineInput[];
}