# lib/api-zod/src/generated/types/lineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LineInputApartmentType } from './lineInputApartmentType';

export interface LineInput {
  /** @minLength 1 */
  payer_name: string;
  /** @minLength 1 */
  payer_email: string;
  /**
     * Required in independent mode
     * @nullable
     */
  apartment_type?: LineInputApartmentType;
  /**
     * Required in independent mode
     * @nullable
     */
  nights?: number | null;
  /**
     * Optional explicit share for split mode (minor units)
     * @nullable
     */
  share_minor?: number | null;
}