# lib/api-zod/src/generated/types/organiserLine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrganiserLine {
  id: string;
  payer_name: string;
  payer_email: string;
  amount_major: number;
  status: string;
  /** @nullable */
  voucher_code?: string | null;
  /** @nullable */
  credit_code?: string | null;
  /** @nullable */
  apartment_type?: string | null;
  /** @nullable */
  nights?: number | null;
  pay_link: string;
}