# lib/api-zod/src/generated/types/organiserView.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrganiserLine } from './organiserLine';

export interface OrganiserView {
  mode: string;
  status: string;
  currency: string;
  organiser_name: string;
  /** @nullable */
  due_by?: string | null;
  /** @nullable */
  split_apartment_type?: string | null;
  /** @nullable */
  split_nights?: number | null;
  /** @nullable */
  split_voucher_code?: string | null;
  paid_count: number;
  total_count: number;
  lines: OrganiserLine[];
}