# lib/api-zod/src/generated/types/rateTable.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RateTableRates } from './rateTableRates';

export interface RateTable {
  currency: string;
  minor_per_major: number;
  rates: RateTableRates;
}