# lib/api-zod/src/generated/types/splitConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SplitConfigApartmentType } from './splitConfigApartmentType';

export interface SplitConfig {
  apartment_type: SplitConfigApartmentType;
  /** @minimum 1 */
  nights: number;
}